refactor: event card examples
