+++ src/index.ts
@@ -7,7 +7,8 @@ export { Badge } from './components/Badge/Badge'
 export { Button } from './components/Button/Button'
 export { ButtonGroup } from './components/ButtonGroup/ButtonGroup'
 export { Calendar } from './components/Calendar/Calendar'
-export { Card } from './components/Card_Old/Card'
+export { Card } from './components/Card/Card'
+export { CardOld } from './components/Card_Old/Card'
 export { CardLoader } from './components/Loaders/CardLoader/CardLoader'
 export { Description } from './components/Description/Description'
 export { DropDown } from './components/DropDown/DropDown'
