+++ src/projects/events/Homepage.stories.tsx
@@ -11,7 +11,7 @@ import { Banner } from '../../_deprecated/Banner/Banner'
 import { TopBar } from '../../components/TopBar/TopBar'
 import { Heading } from '../../components/Heading/Heading'
 import { FooterBasic } from '../../components/Footer/FooterBasic/FooterBasic'
-import { Card } from '../../components/Card_Old/Card'
+import { CardOld as Card } from '../../components/Card_Old/Card'
 import { Listing } from '../../components/Listing/Listing'
 import { Pagination } from '../../components/Pagination/Pagination'
 import { Calendar } from '../../components/Calendar/Calendar'
